refactor(rector): raise the dead-code level to its maximum
Rector was installed and configured with every level at 0, which is to say
with almost no rule enabled: it had never run and had nothing to say. This
takes the dead-code dimension to 68, its maximum, in three passes.

Two rules are skipped rather than obeyed. RemoveNullArgOnNullDefaultParam and
its named-argument twin drop a `null` the signature already defaults to —
true, and exactly wrong here: half the suite passes that null to test it.
`->set('gender')` no longer says which value is set, a test named "handles
null values properly" loses its null values, and the comment "null, not [] —
passing [] there would strip them" in UserObserverTest ends up commenting an
argument that is gone.

Two findings came out of it, both real and both fixed by hand because Rector
could only leave a bare expression behind:

- SendPayementInvite built a GeneratePaymentQR and threw it away. The class
  takes no constructor argument and produces its QR in __invoke(), so the
  line generated nothing; the QR the email carries is built by
  PaymentInvitationEmail itself. The comment above it claimed otherwise.
- rooms/show read auth()->user() into a variable nobody used.

phpstan-baseline.neon drops from 100 entries to 93: seven of them described
errors this pass removed. That is the direction phpstan.neon asks for —
shrink it by fixing entries, never by regenerating it.

public/ leaves withPaths(): it holds index.php and built assets.

+++ app/Domains/Bar/Services/BarCartService.php
@@ -182,7 +182,7 @@ class BarCartService
         return [
             'items' => $items,
             'totalPrice' => (int) $items->sum('total_price'),
-            'cartCount' => (int) array_sum($cart),
+            'cartCount' => array_sum($cart),
         ];
     }
 
@@ -194,7 +194,7 @@ class BarCartService
             return;
         }
 
-        $cart[$productId] = max(0, (int) $cart[$productId] - 1);
+        $cart[$productId] = max(0, $cart[$productId] - 1);
 
         if ($cart[$productId] <= 0) {
             unset($cart[$productId]);
